Clean up IssueCard component
